Move components that should be owned by auth into auth
