+++ apps/browser/src/auth/popup/components/set-pin.component.html
