+++ apps/desktop/src/app/accounts/settings.component.ts
@@ -19,9 +19,9 @@ import { PlatformUtilsService } from "@bitwarden/common/platform/abstractions/pl
 import { Utils } from "@bitwarden/common/platform/misc/utils";
 import { DialogService } from "@bitwarden/components";
 
+import { SetPinComponent } from "../../auth/components/set-pin.component";
 import { flagEnabled } from "../../platform/flags";
 import { ElectronStateService } from "../../platform/services/electron-state.service.abstraction";
-import { SetPinComponent } from "../components/set-pin.component";
 @Component({
   selector: "app-settings",
   templateUrl: "settings.component.html",
