+++ apps/desktop/src/auth/components/set-pin.component.html
